+++ src/test/groovy/org/codenarc/report/AbstractCompactTextReportWriterTestCase.groovy
@@ -15,18 +15,17 @@
  */
 package org.codenarc.report
 
+import static org.codenarc.test.TestUtil.captureSystemOut
+import static org.codenarc.test.TestUtil.shouldFailWithMessageContaining
+import static org.junit.Assert.assertEquals
+
 import org.codenarc.AnalysisContext
 import org.codenarc.results.DirectoryResults
 import org.codenarc.results.FileResults
 import org.codenarc.rule.StubRule
 import org.codenarc.rule.Violation
-import org.codenarc.test.AbstractTestCase
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
-
-import static org.codenarc.test.TestUtil.captureSystemOut
-import static org.codenarc.test.TestUtil.shouldFailWithMessageContaining
-import static org.junit.Assert.assertEquals
 
 /**
  * Abstract superclass for TestReportWriter and subclass tests