+++ src/test/groovy/org/codenarc/report/XmlReportWriterTest.groovy
@@ -15,18 +15,16 @@
  */
 package org.codenarc.report
 
+import static org.codenarc.test.TestUtil.shouldFailWithMessageContaining
+
 import org.codenarc.AnalysisContext
 import org.codenarc.results.DirectoryResults
 import org.codenarc.results.FileResults
-import org.codenarc.rule.StubRule
-import org.codenarc.rule.Violation
 import org.codenarc.rule.imports.DuplicateImportRule
 import org.codenarc.rule.unnecessary.UnnecessaryBooleanInstantiationRule
 import org.codenarc.ruleset.ListRuleSet
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
-
-import static org.codenarc.test.TestUtil.shouldFailWithMessageContaining
 
 /**
  * Tests for XmlReportWriter