+++ CHANGELOG.md
@@ -7,11 +7,14 @@ TODO: Version 3.4.0    (??? 202?)
 --------------------------------------
 Updated/Enhanced Rules and Bug Fixes
  - #751: **ClassStartsWithBlankLine** rule: Fix for interfaces, enums and traits with annotations on the same line. ([Sven](https://github.com/s-ven))
- - #754: Command-line (FilesystemSourceAnalyzer): Match includes and excludes against full source file path and also against path relative to the base directory.
- - #756: Command-line (FilesystemSourceAnalyzer): Fix relative paths on Windows. ([Steven Hartland](https://github.com/stevenh))
  - #739: **NoScriptBindings** rule: Fix false positive when local variable is reassigned.
  - #745: **TrailingComma** rule: Do not require trailing comma if last item is followed by closing bracket on the same line for Maps (that is consistent with Lists).
  - #743: **SpaceAfterCommentDelimiter** and **SpaceBeforeCommentDelimiter** rules: Fix false negatives for dollar-slashy strings.
+ - #748: **UnnecessarySetter** rule: Skip calls to static methods in the same class.
+
+Command-line
+- #754: Command-line (FilesystemSourceAnalyzer): Match includes and excludes against full source file path and also against path relative to the base directory.
+- #756: Command-line (FilesystemSourceAnalyzer): Fix relative paths on Windows. ([Steven Hartland](https://github.com/stevenh))
 
 Documentation
  - #744: Fix wording and typos. ([René Scheibe](https://github.com/darxriggs))